replace trigger changes in filter page
