+++ docs/app/(examples)/examples/filter/page.tsx
@@ -1,30 +1,46 @@
 'use client';
 
-import { Headline, Inset, Panel, Stack, Text } from '@marigold/components';
+import { Button, Inset, Panel, Stack } from '@marigold/components';
+import { Add } from '@marigold/icons';
 import { AppliedFilter } from './applied-filter';
+import { QueryProvider } from './query-provider';
 import { Toolbar } from './toolbar';
-import { VenuesView } from './venues-view';
+import { VenuesTable } from './venues-table';
 
 const FilterPage = () => (
-  <Inset space={4}>
-    <Stack space={8}>
-      <Stack space={2}>
-        <Headline level={2}>Venues</Headline>
-        <Text>Browse and filter available venues for your events.</Text>
+  <QueryProvider>
+    <Inset space={4}>
+      <Stack space={8}>
+        <Panel aria-label="Venues">
+          <Panel.Header>
+            <Panel.Title>Venues</Panel.Title>
+            <Panel.Description>
+              Browse, filter, and manage all venues available for your events.
+            </Panel.Description>
+            <Panel.HeaderActions>
+              <Button
+                variant="primary"
+                onPress={() => {
+                  /* TODO: open Add Venue dialog (DST-1288) */
+                }}
+              >
+                <Add /> Add Venue
+              </Button>
+            </Panel.HeaderActions>
+          </Panel.Header>
+          <Panel.Content>
+            <Stack space="regular">
+              <Toolbar />
+              <AppliedFilter />
+            </Stack>
+          </Panel.Content>
+          <Panel.Content bleed>
+            <VenuesTable />
+          </Panel.Content>
+        </Panel>
       </Stack>
-      <Panel aria-label="Venues">
-        <Panel.Content>
-          <Stack space="regular">
-            <Toolbar />
-            <AppliedFilter />
-          </Stack>
-        </Panel.Content>
-        <Panel.Content bleed>
-          <VenuesView />
-        </Panel.Content>
-      </Panel>
-    </Stack>
-  </Inset>
+    </Inset>
+  </QueryProvider>
 );
 
 export default FilterPage;
