+++ docs/app/(examples)/examples/filter/query-provider.tsx
