+++ docs/app/(examples)/examples/filter/venues-table.tsx
