+++ docs/app/(examples)/examples/navigation.tsx
@@ -64,11 +64,6 @@ const userInput: NavSection = {
       docsHref: '/patterns/user-input/filter',
       docsLabel: 'Form Guidelines',
     },
-    {
-      kind: 'Item',
-      slug: 'data-management',
-      label: 'Data Management',
-    },
   ],
 };
